Move `FolderList` to a separate component
We already do this for resources, so this is nicer :^)
