+++ src/containers/MyNdla/Folders/FoldersPage.tsx
@@ -10,16 +10,12 @@ import { isEqual } from 'lodash';
 import { css } from '@emotion/react';
 import styled from '@emotion/styled';
 import { AddButton } from '@ndla/button';
-import { breakpoints, colors, mq, spacing } from '@ndla/core';
-import { FolderOutlined } from '@ndla/icons/contentType';
-import { Folder, useSnack } from '@ndla/ui';
-import { Pencil } from '@ndla/icons/action';
+import { breakpoints, mq, spacing } from '@ndla/core';
+import { useSnack } from '@ndla/ui';
 import { useEffect, useMemo, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useNavigate, useParams } from 'react-router-dom';
-import { DeleteForever } from '@ndla/icons/editor';
 import { HelmetWithTracker } from '@ndla/tracker';
-import { Spinner } from '@ndla/icons';
 import { GQLFolder, GQLFoldersPageQuery } from '../../../graphqlTypes';
 import { useGraphQuery } from '../../../util/runQueries';
 import ListViewOptions from './ListViewOptions';
@@ -31,15 +27,10 @@ import {
   useUpdateFolderMutation,
 } from '../folderMutations';
 import ResourceList from './ResourceList';
-import {
-  FolderTotalCount,
-  getTotalCountForFolder,
-} from '../../../util/folderHelpers';
 import DeleteModal from '../components/DeleteModal';
-import NewFolder from '../../../components/MyNdla/NewFolder';
-import WhileLoading from '../../../components/WhileLoading';
 import FoldersPageTitle from './FoldersPageTitle';
 import FolderAndResourceCount from './FolderAndResourceCount';
+import FolderList from './FolderList';
 
 interface BlockWrapperProps {
   type?: string;
@@ -49,16 +40,6 @@ const FoldersPageContainer = styled.div`
   display: flex;
   flex-direction: column;
   gap: ${spacing.xsmall};
-`;
-
-const StyledFolderIcon = styled.span`
-  display: flex;
-  padding: ${spacing.small};
-  svg {
-    color: ${colors.brand.primary};
-    height: 20px;
-    width: 20px;
-  }
 `;
 
 export const BlockWrapper = styled.ul<BlockWrapperProps>`
@@ -170,15 +151,6 @@ const FoldersPage = () => {
     }
   }, [folders, focusId, previousFolders]);
 
-  const foldersCount = useMemo(
-    () =>
-      folders?.reduce<Record<string, FolderTotalCount>>((acc, curr) => {
-        acc[curr.id] = getTotalCountForFolder(curr);
-        return acc;
-      }, {}),
-    [folders],
-  );
-
   const {
     updateFolder,
     loading: updateFolderLoading,
@@ -250,55 +222,16 @@ const FoldersPage = () => {
         )}
         <ListViewOptions type={type} onTypeChange={setType} />
       </StyledRow>
-      <WhileLoading isLoading={loading} fallback={<Spinner />}>
-        {folders && (
-          <BlockWrapper type={type}>
-            {isAdding && (
-              <NewFolder
-                icon={
-                  <StyledFolderIcon>
-                    <FolderOutlined />
-                  </StyledFolderIcon>
-                }
-                parentId={folderId ?? 'folders'}
-                onClose={() => setIsAdding(false)}
-                onCreate={onFolderAdd}
-              />
-            )}
-            {folders.map((folder, index) => (
-              <ListItem
-                key={`folder-${index}`}
-                id={`folder-${folder.id}`}
-                tabIndex={-1}>
-                <Folder
-                  key={folder.id}
-                  id={folder.id}
-                  link={`/minndla/folders/${folder.id}`}
-                  title={folder.name}
-                  type={type}
-                  subFolders={foldersCount[folder.id]?.folders}
-                  subResources={foldersCount[folder.id]?.resources}
-                  menuItems={[
-                    {
-                      icon: <Pencil />,
-                      text: t('myNdla.folder.edit'),
-                      onClick: () =>
-                        setFolderAction({ action: 'edit', folder, index }),
-                    },
-                    {
-                      icon: <DeleteForever />,
-                      text: t('myNdla.folder.delete'),
-                      onClick: () =>
-                        setFolderAction({ action: 'delete', folder, index }),
-                      type: 'danger',
-                    },
-                  ]}
-                />
-              </ListItem>
-            ))}
-          </BlockWrapper>
-        )}
-      </WhileLoading>
+      <FolderList
+        onFolderAdd={onFolderAdd}
+        isAdding={isAdding}
+        setIsAdding={setIsAdding}
+        type={type}
+        folders={folders}
+        loading={loading}
+        folderId={folderId}
+        setFolderAction={setFolderAction}
+      />
       {selectedFolder && (
         <ResourceList
           selectedFolder={selectedFolder}
